LOGIN REDIRECT TO HOMEPAGE
Used to redirect customers to account page after logging in. Changed it
up so that customers are now directed to homepage after logging in...
Also, a previous change when hitting cart, customers are taken to cart
page and no longer to the cart pop-up. The problem with cart pop-up was
that the page would scroll back to the top every time you update cart
quantities. Cart page allows customers to adjust cart quantities and
doesn't refresh until the customer hits refresh.