[ BD-11 ] & [ BD-14 ]
BD-11 As a customer I need the left filter nav and cart summary bar to
have a non-transparent background so that on scroll I can still view it
once it covers the lower promo zone images.

BD-14 As a customer I need the navigation div layer to be higher than
the add to cart so that I can view the sale menu when the page is
scrolled down some.